Moved corename in valid
